requirements.txt

click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
python-dotenv==0.17.1
six==1.16.0
Werkzeug==2.0.1
aiohttp==3.8.3
aiosignal==1.3.1
aniso8601==9.0.1
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.1.0
async-timeout==4.0.2
attrs==22.1.0
awsebcli==3.20.3
backcall==0.2.0
banana-dev==3.0.1
bcrypt==4.0.1
beautifulsoup4==4.11.1
bleach==5.0.1
blessed==1.20.0
botocore==1.23.54
cached-property==1.5.2
cement==2.8.2
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.7
click==8.1.3
colorama==0.4.3
cryptography==39.0.1
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
Flask==2.2.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
frozenlist==1.3.3
future==0.16.0
gunicorn==20.1.0
idna==3.3
install==1.3.5
ipykernel==6.17.1
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.23.2
jupyter_client==7.4.7
jupyter_core==5.0.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
llvmlite==0.39.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
multidict==6.0.4
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.4
numpy==1.23.5
openai==0.26.2
packaging==21.3
pandas==1.5.3
pandocfilters==1.5.0
paramiko==3.0.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.4
prometheus-client==0.15.0
prompt-toolkit==3.0.32
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
PyNaCl==1.5.0
pynndescent==0.5.8
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==5.4.1
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
requests==2.26.0
scikit-learn==1.2.1
scipy==1.10.0
semantic-version==2.8.5
Send2Trash==1.8.0
six==1.14.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.1
termcolor==1.1.0
terminado==0.17.0
texttable==1.6.7
threadpoolctl==3.1.0
tinycss2==1.2.1
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
umap-learn==0.5.3
urllib3==1.26.7
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.59.0
Werkzeug==2.2.2
widgetsnbextension==4.0.3
yarl==1.8.2
